backend/aws-python-flask-dynamodb-api/requirements.txt
Flask==1.1.2
Flask-Cors==3.0.10
google-auth==1.29.0
requests==2.25.1
